+++ Assets/VRM10_Samples/ClothSample/ClothWarp/Runtime/SphereTriangle/SphereCapsuleCollision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 990d1e921d604e348ab7fd04216b01d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
